requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
djoser==2.2.0
